Rename folder Aux to Aux_ for windows compatibility
