+++ mlir/include/qcc/Dialect/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(Aux)
+add_subdirectory(Aux_) # NOTE: Aux not allowed on windows
 add_subdirectory(Jasp)
